Refactor: Move AppPreferencesService to service package
