+++ data/src/main/java/com/london/data/repository/list/CustomMovieListRepositoryImpl.kt
@@ -14,7 +14,7 @@ import com.london.data.utils.orZero
 import com.london.domain.entity.movie.Movie
 import com.london.domain.entity.movie.MovieList
 import com.london.domain.entity.shared.PagedFetchResponse
-import com.london.domain.repository.AppPreferencesService
+import com.london.domain.service.AppPreferencesService
 import com.london.domain.repository.CustomMovieListRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.sync.Mutex
