+++ data/src/test/java/com/london/data/local/preference/AppPreferencesServiceImplTest.kt
@@ -4,7 +4,7 @@ import android.content.SharedPreferences
 import com.london.domain.entity.contentrestriction.ContentRestrictionLevel
 import com.london.domain.entity.language.AppLanguage
 import com.london.domain.entity.theme.AppTheme
-import com.london.domain.repository.AppPreferencesService
+import com.london.domain.service.AppPreferencesService
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
