+++ presentation/src/main/java/com/london/presentation/feature/welcome/onboarding/OnboardingViewModel.kt
@@ -5,7 +5,7 @@ import androidx.compose.animation.core.FastOutSlowInEasing
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.pager.PagerState
 import androidx.lifecycle.viewModelScope
-import com.london.domain.repository.AppPreferencesService
+import com.london.domain.service.AppPreferencesService
 import com.london.presentation.shared.base.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineScope
