+++ presentation/src/main/java/com/london/presentation/shared/ContentRestrictionProvider.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import com.london.domain.entity.contentrestriction.ContentRestrictionLevel
-import com.london.domain.repository.AppPreferencesService
+import com.london.domain.service.AppPreferencesService
 
 @Composable
 fun ContentRestrictionProvider(
